+++ JavaDesignPatterns/src/designpatterns/strategy/Strategy.java
@@ -1,5 +1,5 @@
 package designpatterns.strategy;
 
 public interface Strategy {
-	int execute(int a, int b); 
+	double execute(int a, int b); 
 }
